Rendering as down the threshold is not working however. GC is working gracefully.
TODO: Fix CL-Bmp

+++ GED_CORE/SanityCheck/Dim2.cs
@@ -1,9 +1,9 @@
 namespace GED.Core.SanityCheck {
     public struct Dim2Vec {
-        public long x, y;
+        public int x, y;
     }
 
     public struct Dim2Sclr {
-        public ulong x, y;
+        public uint x, y;
     }
 }
